+++ packages/renderer/src/components/basic/ButtonComponent.vue
@@ -76,6 +76,7 @@ function handleClick(): void {
       v-if="childComponentId"
       :surface-id="surfaceId"
       :component-id="childComponentId"
+      :base-path="ctx.dataContext.basePath"
     />
   </button>
 </template>